+++ oshi-core/src/main/java/oshi/software/os/windows/WindowsOSProcess.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020-2024 The OSHI Project Contributors
+ * Copyright 2020-2025 The OSHI Project Contributors
  * SPDX-License-Identifier: MIT
  */
 package oshi.software.os.windows;
@@ -272,10 +272,10 @@ public class WindowsOSProcess extends AbstractOSProcess {
     @Override
     public List<OSThread> getThreadDetails() {
         Map<Integer, ThreadPerformanceData.PerfCounterBlock> threads = tcb == null
-                ? ThreadPerformanceData.buildThreadMapFromPerfCounters(Collections.singleton(this.getProcessID()),
-                        this.getName(), -1)
+                ? queryMatchingThreads(Collections.singleton(this.getProcessID()))
                 : tcb;
         return threads.entrySet().stream().parallel()
+                .filter(entry -> entry.getValue().getOwningProcessID() == this.getProcessID())
                 .map(entry -> new WindowsOSThread(getProcessID(), entry.getKey(), this.name, entry.getValue()))
                 .collect(Collectors.toList());
     }
@@ -291,11 +291,7 @@ public class WindowsOSProcess extends AbstractOSProcess {
             pcb = ProcessPerformanceData.buildProcessMapFromPerfCounters(pids);
         }
         if (USE_PROCSTATE_SUSPENDED) {
-            this.tcb = ThreadPerformanceData.buildThreadMapFromRegistry(null);
-            // otherwise performance counters with WMI backup
-            if (this.tcb == null) {
-                this.tcb = ThreadPerformanceData.buildThreadMapFromPerfCounters(null);
-            }
+            this.tcb = queryMatchingThreads(pids);
         }
         Map<Integer, WtsInfo> wts = ProcessWtsData.queryProcessWtsMap(pids);
         return updateAttributes(pcb.get(this.getProcessID()), wts.get(this.getProcessID()));
@@ -364,6 +360,17 @@ public class WindowsOSProcess extends AbstractOSProcess {
         }
 
         return !this.state.equals(INVALID);
+    }
+
+    private Map<Integer, ThreadPerformanceData.PerfCounterBlock> queryMatchingThreads(Set<Integer> pids) {
+        // fetch from registry
+        Map<Integer, ThreadPerformanceData.PerfCounterBlock> threads = ThreadPerformanceData
+                .buildThreadMapFromRegistry(pids);
+        // otherwise performance counters with WMI backup
+        if (threads == null) {
+            threads = ThreadPerformanceData.buildThreadMapFromPerfCounters(pids, this.getName(), -1);
+        }
+        return threads;
     }
 
     private String queryCommandLine() {